pyproject.toml
[project]
name = "mini-blogger-app-flask"
version = "0.1.0"
description = ""
authors = [
    {name = "Soumya2024",email = "habra4561@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask (>=3.1.0,<4.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "werkzeug (>=3.1.3,<4.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "flask-migrate (>=4.1.0,<5.0.0)",
    "config (>=0.5.1,<0.6.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
